Clean.
