+++ sources/mysql/streaming/iterator.go
@@ -10,27 +10,13 @@ import (
 	"github.com/artie-labs/transfer/lib/typing"
 	"github.com/go-mysql-org/go-mysql/replication"
 
-	"github.com/artie-labs/reader/config"
 	"github.com/artie-labs/reader/lib"
 	"github.com/artie-labs/reader/lib/debezium/transformer"
 	"github.com/artie-labs/reader/lib/maputil"
-	"github.com/artie-labs/reader/lib/storage/persistedmap"
 	"github.com/artie-labs/reader/sources/mysql/adapter"
 )
 
 const offsetKey = "offset"
-
-type Iterator struct {
-	batchSize             int32
-	offsets               *persistedmap.PersistedMap[Position]
-	position              Position
-	includedTablesAdapter map[string]*maputil.MostRecentMap[adapter.Table]
-	streamer              *replication.BinlogStreamer
-	syncer                *replication.BinlogSyncer
-
-	// TODO
-	schemaHistory *persistedmap.PersistedMap[any]
-}
 
 func (i *Iterator) Close() error {
 	i.syncer.Close()
@@ -164,43 +150,4 @@ func (i *Iterator) HasNext() bool {
 func (i *Iterator) CommitOffset() {
 	slog.Info("Committing offset", slog.String("position", i.position.String()))
 	i.offsets.Set(offsetKey, i.position)
-}
-
-func BuildStreamingIterator(cfg config.MySQL) (*Iterator, error) {
-	var pos Position
-	offsets := persistedmap.NewPersistedMap[Position](cfg.StreamingSettings.OffsetFile)
-	if _pos, isOk := offsets.Get(offsetKey); isOk {
-		slog.Info("Found offsets", slog.String("offset", _pos.String()))
-		pos = _pos
-	}
-
-	includedTablesAdapter, err := BuildTablesAdapter(cfg)
-	if err != nil {
-		return nil, fmt.Errorf("failed to build schema history: %w", err)
-	}
-
-	syncer := replication.NewBinlogSyncer(
-		replication.BinlogSyncerConfig{
-			ServerID: cfg.StreamingSettings.ServerID,
-			Flavor:   "mysql",
-			Host:     cfg.Host,
-			Port:     uint16(cfg.Port),
-			User:     cfg.Username,
-			Password: cfg.Password,
-		},
-	)
-
-	streamer, err := syncer.StartSync(pos.ToMySQLPosition())
-	if err != nil {
-		return nil, fmt.Errorf("failed to start sync: %w", err)
-	}
-
-	return &Iterator{
-		batchSize:             cfg.GetStreamingBatchSize(),
-		position:              pos,
-		syncer:                syncer,
-		streamer:              streamer,
-		includedTablesAdapter: includedTablesAdapter,
-		offsets:               offsets,
-	}, nil
 }
